Template controller fix and add block controller
