+++ resources/views/app/panel/templates/project_1.blade.php
@@ -15,10 +15,10 @@
             @endforeach
         </div>
         <div class="flex">
-            <button><a href="{{route('panelblock.create', ['template' => $template->id, 'type' => 'paragraph'])}}">Paragraph</a></button>
-            <button><a href="{{route('panelblock.create', ['template' => $template->id, 'type' => 'image-paragraph'])}}">Image-paragraph</a></button>
-            <button><a href="{{route('panelblock.create', ['template' => $template->id, 'type' => 'paragraph-image'])}}">Paragraph-image</a></button>
-            <button><a href="{{route('panelblock.create', ['template' => $template->id, 'type' => 'gallery'])}}">Gallery</a></button>
+            <button><a href="{{route('panelblock.create', ['template' => $block->template_id, 'type' => 'paragraph'])}}">Paragraph</a></button>
+            <button><a href="{{route('panelblock.create', ['template' => $block->template_id, 'type' => 'image-paragraph'])}}">Image-paragraph</a></button>
+            <button><a href="{{route('panelblock.create', ['template' => $block->template_id, 'type' => 'paragraph-image'])}}">Paragraph-image</a></button>
+            <button><a href="{{route('panelblock.create', ['template' => $block->template_id, 'type' => 'gallery'])}}">Gallery</a></button>
         </div>
     </div>
 </x-page.sidebar>
